refactor(project): Remove storage crate
