+++ crates/browser-core/Cargo.toml
@@ -20,7 +20,6 @@ io.workspace = true
 postcard.workspace = true
 rand.workspace = true
 serde.workspace = true
-storage.workspace = true
 thiserror.workspace = true
 tokio.workspace = true
 tracing.workspace = true
