+++ crates/browser-preferences/Cargo.toml
@@ -9,6 +9,5 @@ browser-args.workspace = true
 clap.workspace = true
 io.workspace = true
 serde.workspace = true
-storage.workspace = true
 toml.workspace = true
 tracing.workspace = true
