+++ crates/cookies/Cargo.toml
@@ -7,8 +7,8 @@ license.workspace = true
 [dependencies]
 database.workspace = true
 rusqlite.workspace = true
-storage.workspace = true
 thiserror.workspace = true
 time.workspace = true
 tracing.workspace = true
 url.workspace = true
+io.workspace = true
