+++ crates/http-cache/Cargo.toml
@@ -8,7 +8,6 @@ version.workspace = true
 bytes.workspace = true
 serde.workspace = true
 sha2.workspace = true
-storage.workspace = true
 database.workspace = true
 http.workspace = true
 url.workspace = true
@@ -18,6 +17,7 @@ rusqlite.workspace = true
 httpdate.workspace = true
 tracing.workspace = true
 http-types.workspace = true
+io.workspace = true
 
 [dev-dependencies]
 serial_test.workspace = true
