+++ crates/http-fetch/Cargo.toml
@@ -16,7 +16,6 @@ http.workspace = true
 manifest.workspace = true
 rand.workspace = true
 reqwest.workspace = true
-storage.workspace = true
 tokio.workspace = true
 tokio-stream.workspace = true
 tokio-util.workspace = true
@@ -25,3 +24,4 @@ thiserror.workspace = true
 tracing.workspace = true
 url.workspace = true
 strum.workspace = true
+io.workspace = true
