+++ crates/io/Cargo.toml
@@ -11,9 +11,9 @@ dirs.workspace = true
 http-types.workspace = true
 rust-embed.workspace = true
 serde.workspace = true
-storage.workspace = true
 strum.workspace = true
 thiserror.workspace = true
+manifest.workspace = true
 tokio.workspace = true
 tracing.workspace = true
 url.workspace = true
