+++ crates/io/src/paths.rs
