--- crates/storage/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "storage"
-version = "0.1.0"
-edition.workspace = true
-license.workspace = true
-
-[dependencies]
-dirs.workspace = true
-manifest.workspace = true
