--- crates/storage/LICENSE
@@ -1,1 +1,0 @@
-/home/andree/Code/browser/LICENSE
